tests.py: Add test to check that 'Service-Unavailable' page shown if notify if down.
